+++ packages/stripe_web/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 5.0.0
+
+**Breaking changes**
+Rename `ApplePayShippingContact` to `PlatformpayshippingContact`.
+
+**Fixes**
+- #1465 Handle nullable result coming from Stripe sdk.
+- #1512 add latest charge field to payment intent (Web only).
+- #1522 add bankidentifiercode to ideal paymentdata.
+
 ## 4.6.0
  - Support Flutter 3.16
  - Fix invalid argument `requires_capture` on web (#1431).